TesMuffler/requirements.txt
<<<<<<< HEAD
wx
flask
flake8
youtube-dl
pyrebase


=======
aiofiles==23.1.0
aiohttp==3.9.1
aiosignal==1.3.1
anyio==3.6.2
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
async-timeout==4.0.3
attrs==23.1.0
automationhat==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.9.3
bidict==0.22.1
bleach==5.0.1
blinker==1.4
blinkt==0.1.2
bottle==0.12.25
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==7.1.2
click-log==0.4.0
colorama==0.4.4
colorzero==1.1
contourpy==1.0.7
cryptography==3.3.2
cupshelpers==1.0
cycler==0.11.0
dataclasses==0.6
dbus-python==1.2.16
deprecation==2.1.0
distro==1.5.0
docutils==0.16
dotty-dict==1.3.1
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
fastapi==0.96.0
fastapi-socketio==0.0.10
Flask==1.1.2
fonttools==4.39.4
fourletterphat==0.1.0
frozenlist==1.4.1
gitdb==4.0.10
GitPython==3.1.29
gotrue==0.5.4
gpiozero==1.6.2
greenlet==3.0.2
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
invoke==1.7.3
isort==5.6.4
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.0
jeepney==0.8.0
Jinja2==3.1.2
keyring==23.13.1
kiwisolver==1.4.4
lazy-object-proxy==0.0.0
libsql-client==0.3.0
logilab-common==1.8.1
lxml==4.6.3
markdown2==2.4.8
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.6.1
microdotphat==0.2.1
more-itertools==9.0.0
mote==0.0.4
motephat==0.0.3
multidict==6.0.4
mypy==0.812
mypy-extensions==0.4.3
nicegui==1.2.17
numpy==1.24.3
oauthlib==3.1.0
obd==0.7.1
orjson==3.9.1
packaging==22.0
pantilthat==0.0.7
paramiko==3.2.0
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
Pint==0.7.2
pkginfo==1.9.2
plotly==5.15.0
postgrest-py==0.10.3
protobuf==4.21.12
proxy-tools==0.1.0
pscript==0.7.7
psutil==5.8.0
pycairo==1.16.2
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
pygame==1.9.6
Pygments==2.15.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyparsing==3.0.9
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-gitlab==3.12.0
python-multipart==0.0.6
python-prctl==1.7
python-semantic-release==7.32.1
python-socketio==5.8.0
pywebview==4.1
PyYAML==6.0
rainbowhat==0.1.0
readme-renderer==37.3
realtime==0.0.5
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
requests-toolbelt==0.10.1
responses==0.12.1
rfc3986==1.5.0
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==3.3.3
semver==2.13.0
Send2Trash==1.6.0b1
sense-hat==2.6.0
simpleaudio==1.0.4
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
smmap==5.0.0
sn3218==1.2.7
sniffio==1.3.0
soupsieve==2.2.1
spidev==3.5
SQLAlchemy==2.0.23
sqlalchemy-libsql==0.1.0
ssh-import-id==5.10
starlette==0.27.0
storage3==0.3.5
supafunc==0.2.2
tenacity==8.2.2
thonny==4.0.1
toml==0.10.1
tomlkit==0.11.6
touchphat==0.0.1
tqdm==4.64.1
twine==3.8.0
twython==3.8.2
typed-ast==1.4.2
typing_extensions==4.9.0
unicornhathd==0.0.4
urllib3==1.26.5
uvicorn==0.20.0
uvloop==0.17.0
v4l2-python3==0.3.2
vbuild==0.8.1
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.9.4
zipp==3.11.0
>>>>>>> b15dd0a69ddf2ae0e3525153edcd12ecbc42638b
